Rocket Game
In Rocket coins game, players must analyse and detect the correct missile. In this system, there are three rockets. Namely; Bitcoin, Ethereum, Litecoin, and they are competing with each other. The prizewinner will be determined according to the player that selects the winning rocket. Prize structure Players win the winning rocket will receive a prize of x1.5 + the original amount  
 
Example: You bet 100 Gae on a Bitcoin rocket and win. You get a
bonus of 150 Gae + 100 Gae = 250 Gae